fix benchmarks on 32-bit platforms

+++ benches/iterbench.rs
@@ -18,7 +18,7 @@ fn iter_sum(bench: &mut Bencher) {
     let img = Img::new_stride(vec![0x11223344u32; 802*600], 800, 600, 802);
 
     bench.iter(|| {
-        img.pixels().map(|p| p as usize).sum::<usize>()
+        img.pixels().map(|p| p as u64).sum::<u64>()
     });
 }
 
@@ -27,7 +27,7 @@ fn stride_ignorant_sum(bench: &mut Bencher) {
     let img = Img::new_stride(vec![0x11223344u32; 802*600], 800, 600, 802);
 
     bench.iter(|| {
-        img.buf().iter().copied().map(|p| p as usize).sum::<usize>()
+        img.buf().iter().copied().map(|p| p as u64).sum::<u64>()
     });
 }
 
@@ -36,6 +36,6 @@ fn chunked_sum(bench: &mut Bencher) {
     let img = Img::new_stride(vec![0x11223344u32; 802*600], 800, 600, 802);
 
     bench.iter(|| {
-        img.buf().chunks(img.stride()).flat_map(|row| row[0..img.width()].iter()).copied().map(|p| p as usize).sum::<usize>()
+        img.buf().chunks(img.stride()).flat_map(|row| row[0..img.width()].iter()).copied().map(|p| p as u64).sum::<u64>()
     });
 }
